+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/registration/RegistrationMessageFilter.java
@@ -13,7 +13,7 @@
 package org.eclipse.hono.service.registration;
 
 import org.apache.qpid.proton.message.Message;
-import org.eclipse.hono.util.BaseMessageFilter;
+import org.eclipse.hono.service.BaseMessageFilter;
 import org.eclipse.hono.util.MessageHelper;
 import org.eclipse.hono.util.ResourceIdentifier;
 import org.slf4j.Logger;